// soy-rappi/src/app/utils/strapi/links/models/link.ts
export interface Link {
  slug: string;
  reference: LinkReference;
  contentType: LinkContentType;
  referenceId: string;
  locale: string;
}

export enum LinkReference {
  ACCEPTING_ORDERS_PAGE = 'ACCEPTING_ORDERS_PAGE',
  BAG_PAGE = 'BAG_PAGE',
  BANK_ACCOUNT_CALENDAR_PAGE = 'BANK_ACCOUNT_CALENDAR_PAGE',
  BLOG_PAGE = 'BLOG_PAGE',
  DISCIPLINARY_PAGE = 'DISCIPLINARY_PAGE',
  EARN_PAGE = 'EARN_PAGE',
  EXCLUSIVE_DISCOUNTS_PAGE = 'EXCLUSIVE_DISCOUNTS_PAGE',
  FIVE_STARS_PAGE = 'FIVE_STARS_PAGE',
  HEALTH_PAGE = 'HEALTH_PAGE',
  HOME_PAGE = 'HOME_PAGE',
  INFORMATIVE_VIDEOS_PAGE = 'INFORMATIVE_VIDEOS_PAGE',
  MONOTAX_PAGE = 'MONOTAX_PAGE',
  ORDER_TYPES_PAGE = 'ORDER_TYPES_PAGE',
  OUR_STORY_PAGE = 'OUR_STORY_PAGE',
  RAPPI_METRICS_PAGE = 'RAPPI_METRICS_PAGE',
  RAPPI_TOOLS_PAGE = 'RAPPI_TOOLS_PAGE',
  REFERENCES_PAGE = 'REFERENCES_PAGE',
  SCHEDULE_RESERVATION_PAGE = 'SCHEDULE_RESERVATION_PAGE',
  SHOPPING_CARD_PAGE = 'SHOPPING_CARD_PAGE',
  STATUS_PAGE = 'STATUS_PAGE',
  TIPS_PAGE = 'TIPS_PAGE',
  WHAT_HAVE_KNOW_PAGE = 'WHAT_HAVE_KNOW_PAGE',
  WHO_WE_ARE_PAGE = 'WHO_WE_ARE_PAGE',
  DYNAMIC_PAGE = 'DYNAMIC_PAGE',
  REWARDS_PAGE = 'REWARDS_PAGE',
  THANK_YOU_PAGE = 'THANK_YOU_PAGE',
}

export enum LinkContentType {
  singleType = 'singleType',
  collectionType = 'collectionType',
}